refactor ray metrics
